requirements.txt
paho-mqtt>=1.6.1
matplotlib>=3.5.0
numpy>=1.21.0 